--- packages/apollo-collaboration-server/src/utils/role/role.decorator.ts
@@ -1,6 +1,0 @@
-import { SetMetadata } from '@nestjs/common'
-
-import { Role } from './role.enum'
-
-export const ROLES_KEY = 'roles'
-export const Roles = (...roles: Role[]) => SetMetadata(ROLES_KEY, roles)
