+++ packages/apollo-collaboration-server/src/utils/validation/validatation.decorator.ts
@@ -1,0 +1,6 @@
+import { SetMetadata } from '@nestjs/common'
+
+import { Role } from '../role/role.enum'
+
+export const ROLES_KEY = 'roles'
+export const Validations = (...roles: Role[]) => SetMetadata(ROLES_KEY, roles)
